+++ client/index.html
@@ -49,9 +49,10 @@
 							</div>
 						</form>
 					</div>
-					<div class="o-forms-group">
+					<div id="tech-radar__form-container" class="o-forms-group collapsed">
 						<label class="o-forms-label">Settings</label>
-						<form  id="tech-radar__qp-form" action=""></form>
+						<form  id="tech-radar__qp-form" action="">
+						</form>
 					</div>
 				</div>
 			</div>
